Typo Fix on top of every files - fix
Top of every files were inconsistent, like it was quite a mess to decode what
headers has been added or not in what order.

I took the time to fix every files and set a top structure:
  - Comments
  * 2 wide spaces
  - Headers included in the projects
  - SDL2 headers
  - additionnal libraries
  * 2 wide spaces
  - codes

This way i found out it much easier to read and understand what's going on.

Also i added an uppercase to the m from the main.* files.
Main.cpp and Main.h is more in addequation with others files of the project.

(i really don't know how to use git holy hell)